+++ lib/src/database/upgrades/heavy_weapon/z_6_trooper.dart
@@ -7,7 +7,7 @@ final z6Trooper = Upgrade(
   id: 'z-6-trooper',
   name: 'Z-6 Trooper',
   type: UpgradeSlot.heavyWeapon,
-  restrictedToUnit: [unit.rebelTroopers],
+  restrictedToUnit: [unit.rebelTroopers.toRef()],
   addsMiniature: true,
   weapon: weapon.z6RotaryBlaster,
   points: 22,